pip_requirements.txt
anyio==3.6.1
astroid==2.12.5
click==8.1.3
dill==0.3.5.1
fastapi==0.81.0
greenlet==1.1.3
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pydantic==1.10.1
pylint==2.15.0
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
uvicorn==0.18.3
wrapt==1.14.1
